NewsProxy/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-celery-beat==2.5.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
kombu==5.3.5
packaging==23.2
prompt-toolkit==3.0.43
python-dateutil==2.8.2
python-memcached==1.62
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
